requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyodbc==5.2.0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
